QuizGame
The title screen should include the title of the game ("My Quiz Game" or similar) and a button to start the game.
The game screen should include the question text, as well as 2 - 4 buttons for different answers.
The ending screen should display how many questions the player got right/wrong, as well as a button for returning to the title screen.
There are many ways you can handle the logic here, but the simplest might be to make different functions such as RightAnswer and WrongAnswer, that are handled by the appropriate buttons. Additionally, you will need a GameManager class to keep track of how many questions have been answered correctly. Be sure to reset the game's score when it is restarted!